Rename things and remove printlns
