+++ extras/queue-manager-replicated/core/pom.xml
