+++ extras/queue-manager-replicated/core/src/main/resources/META-INF/beans.xml
