+++ extras/queue-manager-replicated/pom.xml
