+++ extras/queue-manager-replicated/replication-mp-reactive/src/main/resources/META-INF/beans.xml
