+++ extras/queue-manager-replicated/tests/pom.xml
