+++ extras/queue-manager-replicated/tests/src/test/resources/application.properties
